requirements.txt
langchain==0.3.26
langchain_community==0.3.26
langchain_core==0.3.67
langchain_google_genai==2.1.6
langchain_text_splitters==0.3.8
python-dotenv==1.1.1
streamlit==1.45.1
sympy==1.14.0
